mudanca no modo de salvar os valores no banco

+++ src/app/models/Freight.js
@@ -9,11 +9,11 @@ class Freight extends Model {
         final_freight_city: Sequelize.STRING, // cidade final do frete
         location_of_the_truck: Sequelize.STRING, // local atual do caminhão
         contractor: Sequelize.STRING, // empresa que foi pego o frete
-        truck_current_km: Sequelize.DOUBLE, // km atual registrado no caminhão
-        liter_of_fuel_per_km: Sequelize.DOUBLE, // media do caminhão
-        preview_tonne: Sequelize.DOUBLE, // previa de tonelada
-        preview_value_diesel: Sequelize.DOUBLE, // previa de valor do combustivel
-        value_tonne: Sequelize.DOUBLE, // valor da tonelada
+        truck_current_km: Sequelize.DECIMAL, // km atual registrado no caminhão
+        liter_of_fuel_per_km: Sequelize.DECIMAL, // media do caminhão
+        preview_tonne: Sequelize.DECIMAL, // previa de tonelada
+        preview_value_diesel: Sequelize.DECIMAL, // previa de valor do combustivel
+        value_tonne: Sequelize.DECIMAL, // valor da tonelada
         status: Sequelize.ENUM({
           values: [
             'APPROVAL_PROCESS',
@@ -24,10 +24,10 @@ class Freight extends Model {
           ],
         }),
         // level two
-        tons_loaded: Sequelize.DOUBLE, // total da tonelada carregada
-        toll_value: Sequelize.DOUBLE, // valor do pedagio
-        truck_km_completed_trip: Sequelize.DOUBLE, // km do caminhão do final da viagem
-        discharge: Sequelize.DOUBLE, // tonelada descarregada
+        tons_loaded: Sequelize.DECIMAL, // total da tonelada carregada
+        toll_value: Sequelize.DECIMAL, // valor do pedagio
+        truck_km_completed_trip: Sequelize.DECIMAL, // km do caminhão do final da viagem
+        discharge: Sequelize.DECIMAL, // tonelada descarregada
         img_proof_cte: Sequelize.STRING, // url imagem cte
         img_proof_ticket: Sequelize.STRING, // url imagem ticket
         img_proof_freight_letter: Sequelize.STRING, // url imagem carta frete
